moved test outside
